Add WordBoundaryFinder.js to the repository
WordBoundaryFinder acts as a wrapper for ICU's BreakIterator class.  WordBoundaryFinder can be configured for a language with configure().  findBoundaries() takes a string and returns an array of indices where each index is a word boundary in the string.  This object depends on icu.data.  initialize() must be called before calling configure() or findBoundaries().