added spellchecker.js to the repo
spellchecker.js serves as a wrapper to spell checking functions provided by Hunspell.  The Spellchecker object has three functions: loadDictionary(), check(), and suggest().  loadDictionary() configures the spellchecker for a given language and locale.  loadDictionary() should be called after adding a .aff and .dic file to the Emscripten file system with FS.createPreloadedFile().
check() takes a word and returns 1 if the word is spelled correctly and 0 if spelled incorrectly.
suggest() takes a word and returns an array of spelling suggestions.